Add tests for detached buffers, and move typed array tests into separate modules
